apps/agent-credit-system/requirements.txt
uvicorn==0.34.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
ipython-sql==0.5.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
regex==2024.11.6
psycopg2-binary==2.9.9
fastapi==0.115.8
urllib3==2.3.0
databases==0.9.0
python-multipart==0.0.20
ulid==1.1
logfire==3.5.3
openai==1.63.2
pytz==2025.1
jinja2==3.1.3
aiofiles==23.2.1
passlib==1.7.4
python-jose==3.3.0
google-cloud-storage==2.16.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
twilio==9.3.7
pydantic==2.6.4
email-validator==2.2.0
phonenumbers==8.13.32
